refactor: back-channel logout is a profile of the SET core, not a Shared Signals feature
The two packages divide by a question. Abblix.SecurityEvents owns the token and the wire - what a
SET is, how it is validated, how one travels. Abblix.SharedSignals owns the relationship - the
stream two parties agreed on, its lifecycle and subjects, and the metadata by which they find each
other. What decides placement is whether a thing stops making sense without a stream.

A logout notification has no stream: one token, delivered once, by a provider the application
already knows from elsewhere. It uses the token and the pipeline, and both are in the core - which
the move proves rather than asserts, because the core's project references are unchanged by it,
and the moved folder's every using already named the core.

It sat in the wrong package because the profile keys did. A published constant drags whatever
names it into its own package, since the code using it has to live where it can see it - so a key
holder named after one framework had made every profile it named that framework's. The keys move
to the core too, as ValidationProfileKeys, and for the reason the holder now states: a profile is
a shape of this core's pipeline, whatever specification asked for that shape.

The idempotent profile registration becomes public here, because both receivers ask the container
the same question and the second copy of a question is what drifts.
