+++ src/Abblix.SecurityEvents/BackChannelLogout/BackChannelLogoutError.cs
@@ -22,7 +22,7 @@
 
 using System.Text.Json.Serialization;
 
-namespace Abblix.SharedSignals.Receiver.BackChannelLogout;
+namespace Abblix.SecurityEvents.BackChannelLogout;
 
 /// <summary>
 /// How an RP tells a provider that a logout request was bad: the optional body of the 400 response
