+++ src/Abblix.SecurityEvents/BackChannelLogout/ILogoutNotificationSink.cs
@@ -20,7 +20,7 @@
 // CONTACT: For license inquiries or permissions, contact Abblix LLP at
 // info@abblix.com
 
-namespace Abblix.SharedSignals.Receiver.BackChannelLogout;
+namespace Abblix.SecurityEvents.BackChannelLogout;
 
 /// <summary>
 /// Where validated logout orders land: the application's half of the receiver, called once per
