+++ src/Abblix.SecurityEvents/BackChannelLogout/ILogoutTokenValidator.cs
@@ -20,7 +20,7 @@
 // CONTACT: For license inquiries or permissions, contact Abblix LLP at
 // info@abblix.com
 
-namespace Abblix.SharedSignals.Receiver.BackChannelLogout;
+namespace Abblix.SecurityEvents.BackChannelLogout;
 
 /// <summary>
 /// Validates the Logout Token a provider posts to this receiver to say a session has ended
